refactor(shared): move Tenant entity from Identity to Shared with Doctrine XML mapping
Tenant is the only true shared kernel — every business BC carries a
tenant_id foreign key. Move the aggregate out of Identity, where it
lived for legacy reasons, and into a new Shared/Domain/Tenant.

The new home uses XML mapping in Shared/Infrastructure/Doctrine/Orm/
Mapping/Tenant.orm.xml so the Domain class no longer carries Doctrine
attributes — first BC to lead the way to the per-BC XML migration in
RF-06..RF-09.

Cross-BC consumers (40+ call sites) keep working untouched: a
class_alias() bridge in src/Identity/Domain/Entity/Tenant.php turns the
legacy FQCN into an alias of the canonical Shared\Domain\Tenant. The
Doctrine resolve_target_entities entry rewires `targetEntity:` strings
the same way at metadata-load time. Both bridges are removed in RF-04
once every consumer is updated.

A repository port is introduced as well: Shared\Domain\Repository\
TenantRepositoryInterface plus DoctrineTenantRepository in Shared/
Infrastructure with the canonical "Doctrine" prefix; the legacy
Identity\Infrastructure\Doctrine\Repository\TenantRepository becomes a
deprecated final subclass for autowiring continuity.

Tooling tweaks: PHPStan now bootstraps the runtime aliases through
tools/phpstan/bootstrap-aliases.php so the static analyzer follows the
same class graph as the runtime.

Refs #152
