+++ apps/api/config/packages/doctrine.yaml
@@ -43,6 +43,12 @@ doctrine:
         # with the bundle-style structure under src/{Context}/Domain/Entity.
         auto_mapping: false
         mappings:
+            Shared:
+                type: xml
+                is_bundle: false
+                dir: '%kernel.project_dir%/src/Shared/Infrastructure/Doctrine/Orm/Mapping'
+                prefix: 'App\Shared\Domain'
+                alias: Shared
             Catalog:
                 type: attribute
                 is_bundle: false
@@ -69,6 +75,14 @@ doctrine:
                 alias: Identity
         controller_resolver:
             auto_mapping: false
+
+        # Bridge for the RF-02 Tenant move: cross-BC entities still declare
+        # `targetEntity: App\Identity\Domain\Entity\Tenant` until RF-04 rewrites
+        # them. Doctrine's targetEntity resolver swaps the legacy FQCN out for
+        # the canonical Shared\Domain\Tenant at metadata-load time so foreign
+        # keys keep pointing at the same physical entity.
+        resolve_target_entities:
+            App\Identity\Domain\Entity\Tenant: App\Shared\Domain\Tenant
 
         # Registered (declared) but disabled by default. Each request enables
         # it via TenantFilterConfigurator once the active tenant is known.
