+++ apps/api/src/Shared/Infrastructure/Doctrine/Orm/Mapping/Tenant.orm.xml
@@ -1,0 +1,32 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<doctrine-mapping xmlns="http://doctrine-project.org/schemas/orm/doctrine-mapping"
+                  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+                  xsi:schemaLocation="http://doctrine-project.org/schemas/orm/doctrine-mapping
+                                      https://www.doctrine-project.org/schemas/orm/doctrine-mapping.xsd">
+
+    <entity name="App\Shared\Domain\Tenant"
+            table="tenants"
+            repository-class="App\Shared\Infrastructure\Doctrine\Repository\DoctrineTenantRepository">
+
+        <unique-constraints>
+            <unique-constraint name="tenants_code_uniq" columns="code"/>
+            <unique-constraint name="tenants_domain_uniq" columns="domain"/>
+        </unique-constraints>
+
+        <id name="id" type="uuid" column="id">
+            <generator strategy="NONE"/>
+        </id>
+
+        <field name="code" type="string" length="64"/>
+        <field name="name" type="string" length="255"/>
+        <field name="domain" type="string" length="255" nullable="true"/>
+        <field name="plan" type="string" length="32">
+            <options>
+                <option name="default">starter</option>
+            </options>
+        </field>
+        <field name="createdAt" type="datetime_immutable" column="created_at"/>
+
+    </entity>
+
+</doctrine-mapping>
